Fixed incorrect output with sample sizes 2^16 and larger

+++ mcpi_const.h
@@ -2,7 +2,7 @@
 #define MCPI_CONST_H_
 
 typedef double vec_t;
-typedef unsigned int_t;
+typedef unsigned long long int_t;
 
 #define BUFFER_SIZE_SQRT pow(2, 16)
 #define NUM_THREADS 4
